Refactoring - added package repository
